refactor(shelf): move optimistic save-queue into useShelfMutations
Tier 3 architecture review (A3).

Moved the non-trivial concurrency/optimistic logic out of the
BookshelfSection component body into a useShelfMutations hook:
- the serialized, latest-wins-per-book save queue (pendingRef/runningRef
  guarding the non-atomic mutateMultiSelect backend),
- the optimistic "przeczytane" overrides with rollback,
- the optimistic order-plan save (applyOrderPlan → saveOrders + rollback),
- the shared moveError.

The component now renders and wires callbacks; the pure helpers
(planInsertion / splitShelves / featuredReads) were already extracted.
236 → 177 lines. No behavior change. Suite 474 green, lint clean, build OK.
