+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-example",
   "private": true,
-  "version": "1.67.7",
+  "version": "1.67.8",
   "type": "module",
   "engines": {
     "node": ">=18 <21"
